update for deployment

+++ app/models/alkuperaa_koskevat_tiedot.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,9 +11,9 @@ class AlkuperaaKoskevatTiedot(Base):
     __tablename__ = 'alkuperaa_koskevat_tiedot'
     __table_args__ = (ForeignKeyConstraint(['hankintaID'], ['hankintatiedot.hankintaID'], name='alkuperaa_koskevat_tiedot_ibfk_1'), Index('IDX_Alkuperaa_koskevat_tiedot1', 'hankintaID'))
     # Origin number
-    alkupera_nro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    alkupera_nro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Acquisition ID
-    hankintaID: Mapped[int] = mapped_column(INTEGER(11), nullable=False, server_default=text('0'))
+    hankintaID: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text('0'))
     # Origin type
     alkuperatyyppi: Mapped[Optional[str]] = mapped_column(String(255))
     # Country
