+++ app/models/kasvatustietoja.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,9 +11,9 @@ class Kasvatustietoja(Base):
     __tablename__ = 'kasvatustietoja'
     __table_args__ = (ForeignKeyConstraint(['hankintaID'], ['hankintatiedot.hankintaID'], name='kasvatustietoja_ibfk_1'), Index('IDX_Kasvatustietoja1', 'hankintaID'))
     # Cultivation additional information number
-    lisatietojen_nro_kasvatus: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    lisatietojen_nro_kasvatus: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Acquisition ID
-    hankintaID: Mapped[int] = mapped_column(INTEGER(11), nullable=False, server_default=text('0'))
+    hankintaID: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text('0'))
     # Seed bank
     siemenpankki: Mapped[Optional[str]] = mapped_column(String(255))
     # Seeds remaining
