+++ app/models/kasvin_kayttotarkoitus.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,9 +11,9 @@ class KasvinKayttotarkoitus(Base):
     __tablename__ = 'kasvin_kayttotarkoitus'
     __table_args__ = (ForeignKeyConstraint(['taksonin_nro'], ['taksoni.taksonin_nro'], name='kasvin_kayttotarkoitus_ibfk_1'), ForeignKeyConstraint(['viitenro'], ['viite.viitenro'], name='kasvin_kayttotarkoitus_ibfk_2'), Index('IDX_Kasvin_kayttotarkoitus1', 'taksonin_nro'), Index('IDX_Kasvin_kayttotarkoitus2', 'viitenro'))
     # Usage number
-    kayttonro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    kayttonro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Taxon number
-    taksonin_nro: Mapped[int] = mapped_column(INTEGER(11), nullable=False, server_default=text('0'))
+    taksonin_nro: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text('0'))
     # Usage code
     kayton_tunnus: Mapped[Optional[str]] = mapped_column(String(255))
     # Usage
@@ -21,7 +21,7 @@ class KasvinKayttotarkoitus(Base):
     # Explanation
     selite: Mapped[Optional[str]] = mapped_column(String(255))
     # Reference number
-    viitenro: Mapped[Optional[int]] = mapped_column(INTEGER(11))
+    viitenro: Mapped[Optional[int]] = mapped_column(Integer)
     # Taxon
     taksoni: Mapped['Taksoni'] = relationship('Taksoni', back_populates='kasvin_kayttotarkoitus')
     # Reference or bibliographic citation
