+++ app/models/koe.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -10,8 +10,8 @@ class Koe(Base):
     # Experimental or test table
     __tablename__ = 'koe'
     # ID
-    id: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Value
-    i: Mapped[Optional[int]] = mapped_column(INTEGER(11))
+    i: Mapped[Optional[int]] = mapped_column(Integer)
     # Text
     t: Mapped[Optional[str]] = mapped_column(String(255))
