+++ app/models/lahettaja.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -10,7 +10,7 @@ class Lahettaja(Base):
     # Sender or source
     __tablename__ = 'lahettaja'
     # Sender number
-    lahettajanro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    lahettajanro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Sender type
     lahettajatyyppi: Mapped[Optional[str]] = mapped_column(String(255))
     # Sender identifier in the garden
