+++ app/models/lista_alkuperatyyppi.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -12,4 +12,4 @@ class ListaAlkuperatyyppi(Base):
     # Name
     nimi: Mapped[str] = mapped_column(String(255), nullable=False, server_default=text("''"))
     # ID
-    id: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
