+++ app/models/lista_lahettajantyyppi.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -10,6 +10,6 @@ class ListaLahettajantyyppi(Base):
     # Sender type list
     __tablename__ = 'lista_lahettajantyyppi'
     # ID
-    id: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Name
     nimi: Mapped[Optional[str]] = mapped_column(String(255))
