+++ app/models/lista_luonnonsuojeluarvo_muualla.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -10,7 +10,7 @@ class ListaLuonnonsuojeluarvoMuualla(Base):
     # Nature conservation value elsewhere list
     __tablename__ = 'lista_luonnonsuojeluarvo_muualla'
     # ID
-    ID: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    ID: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Name
     NIMI: Mapped[Optional[str]] = mapped_column(String(255))
     # Code
