+++ app/models/maailman_levinneisyysalue.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,9 +11,9 @@ class MaailmanLevinneisyysalue(Base):
     __tablename__ = 'maailman_levinneisyysalue'
     __table_args__ = (ForeignKeyConstraint(['taksonin_nro'], ['taksoni.taksonin_nro'], name='maailman_levinneisyysalue_ibfk_1'), Index('IDX_Maailman_Levinneisyysalue1', 'taksonin_nro'))
     # World distribution area number
-    levinneisyysalueen_nro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    levinneisyysalueen_nro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Taxon number
-    taksonin_nro: Mapped[int] = mapped_column(INTEGER(11), nullable=False, server_default=text('0'))
+    taksonin_nro: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text('0'))
     # Distribution area
     levinneisyysalue: Mapped[Optional[str]] = mapped_column(String(255))
     # Distribution area detail
@@ -21,7 +21,7 @@ class MaailmanLevinneisyysalue(Base):
     # Origin or introduced status
     alkuperainen_vai_tulokas: Mapped[Optional[str]] = mapped_column(String(255))
     # Reference number
-    viitenro: Mapped[Optional[int]] = mapped_column(INTEGER(11))
+    viitenro: Mapped[Optional[int]] = mapped_column(Integer)
     # Additional information
     lisatietoja: Mapped[Optional[str]] = mapped_column(String(255))
     # Taxon
