+++ app/models/maaritysmerkinta.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,7 +11,7 @@ class Maaritysmerkinta(Base):
     __tablename__ = 'maaritysmerkinta'
     __table_args__ = (ForeignKeyConstraint(['hankintaID'], ['hankintatiedot.hankintaID'], name='maaritysmerkinta_ibfk_1'), Index('IDX_Maaritysmerkinta1', 'hankintaID'))
     # Maaritysnro
-    maaritysnro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    maaritysnro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Identification date
     maarityspvm: Mapped[Optional[str]] = mapped_column(String(255))
     # Identifier
@@ -21,7 +21,7 @@ class Maaritysmerkinta(Base):
     # Remark
     huomautus: Mapped[Optional[str]] = mapped_column(String(255))
     # Acquisition ID
-    hankintaID: Mapped[Optional[int]] = mapped_column(INTEGER(11))
+    hankintaID: Mapped[Optional[int]] = mapped_column(Integer)
     # Section
     osasto: Mapped[Optional[str]] = mapped_column(String(255))
     # Paikka
