+++ app/models/muunkielinen_nimi.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,17 +11,17 @@ class MuunkielinenNimi(Base):
     __tablename__ = 'muunkielinen_nimi'
     __table_args__ = (ForeignKeyConstraint(['taksonin_nro'], ['taksoni.taksonin_nro'], name='muunkielinen_nimi_ibfk_1'), ForeignKeyConstraint(['viitenro'], ['viite.viitenro'], name='muunkielinen_nimi_ibfk_2'), Index('IDX_Muunkielinen_nimi1', 'taksonin_nro'), Index('IDX_Muunkielinen_nimi2', 'viitenro'))
     # Nimen number
-    nimen_nro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    nimen_nro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Taxon number
-    taksonin_nro: Mapped[int] = mapped_column(INTEGER(11), nullable=False, server_default=text('0'))
+    taksonin_nro: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text('0'))
     # Name
     nimi: Mapped[Optional[str]] = mapped_column(String(255))
     # Kieli
     kieli: Mapped[Optional[str]] = mapped_column(String(255))
     # Reference number
-    viitenro: Mapped[Optional[int]] = mapped_column(INTEGER(11))
+    viitenro: Mapped[Optional[int]] = mapped_column(Integer)
     # Reference or bibliographic citation 2
-    viite_2: Mapped[Optional[int]] = mapped_column(INTEGER(11))
+    viite_2: Mapped[Optional[int]] = mapped_column(Integer)
     # Taxon
     taksoni: Mapped['Taksoni'] = relationship('Taksoni', back_populates='muunkielinen_nimi')
     # Reference or bibliographic citation
