+++ app/models/naytetieto.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,9 +11,9 @@ class Naytetieto(Base):
     __tablename__ = 'naytetieto'
     __table_args__ = (ForeignKeyConstraint(['taksonin_nro'], ['taksoni.taksonin_nro'], name='naytetieto_ibfk_1'), ForeignKeyConstraint(['viitenro'], ['viite.viitenro'], name='naytetieto_ibfk_2'), Index('IDX_Naytetieto1', 'taksonin_nro'), Index('IDX_Naytetieto2', 'viitenro'))
     # Specimen number
-    naytteen_nro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    naytteen_nro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Taxon number
-    taksonin_nro: Mapped[int] = mapped_column(INTEGER(11), nullable=False, server_default=text('0'))
+    taksonin_nro: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text('0'))
     # Type
     tyyppi: Mapped[Optional[str]] = mapped_column(String(255))
     # Location
@@ -25,7 +25,7 @@ class Naytetieto(Base):
     # Date
     paivays: Mapped[Optional[str]] = mapped_column(String(255))
     # Reference number
-    viitenro: Mapped[Optional[int]] = mapped_column(INTEGER(11))
+    viitenro: Mapped[Optional[int]] = mapped_column(Integer)
     # Location explanation
     sijainnin_selite: Mapped[Optional[str]] = mapped_column(String(255))
     # Reference explanation
