+++ app/models/naytetietoja.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,7 +11,7 @@ class Naytetietoja(Base):
     __tablename__ = 'naytetietoja'
     __table_args__ = (ForeignKeyConstraint(['hankintaID'], ['hankintatiedot.hankintaID'], name='naytetietoja_ibfk_1'), Index('IDX_Naytetietoja1', 'hankintaID'))
     # Specimen number
-    naytteen_nro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    naytteen_nro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Specimen type
     naytteen_tyyppi: Mapped[Optional[str]] = mapped_column(String(255))
     # Specimen location
@@ -23,11 +23,11 @@ class Naytetietoja(Base):
     # Specimen date
     naytteen_paivays: Mapped[Optional[str]] = mapped_column(String(255))
     # Acquisition ID
-    hankintaID: Mapped[Optional[int]] = mapped_column(INTEGER(11))
+    hankintaID: Mapped[Optional[int]] = mapped_column(Integer)
     # Location explanation
     sijainnin_selite: Mapped[Optional[str]] = mapped_column(String(255))
     # Reference number
-    viitenro: Mapped[Optional[int]] = mapped_column(INTEGER(11))
+    viitenro: Mapped[Optional[int]] = mapped_column(Integer)
     # Reference explanation
     viitteen_selite: Mapped[Optional[str]] = mapped_column(String(255))
     # New specimen date
