+++ app/models/puutarhassa_viljelyn_tarkoitus.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,9 +11,9 @@ class PuutarhassaViljelynTarkoitus(Base):
     __tablename__ = 'puutarhassa_viljelyn_tarkoitus'
     __table_args__ = (ForeignKeyConstraint(['hankintaID'], ['hankintatiedot.hankintaID'], name='puutarhassa_viljelyn_tarkoitus_ibfk_1'), Index('IDX_Puutarhassa_viljelyn_tarkoitus1', 'hankintaID'))
     # Viljely number
-    viljely_nro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    viljely_nro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Acquisition ID
-    hankintaID: Mapped[int] = mapped_column(INTEGER(11), nullable=False, server_default=text('0'))
+    hankintaID: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text('0'))
     # Garden cultivation purpose
     puutarhassa_viljelyn_tarkoitus: Mapped[Optional[str]] = mapped_column(String(255))
     # In the garden cultivation detail
