+++ app/models/sijoituspaikka.py
@@ -2,7 +2,7 @@ from .base import Base
 from typing import Optional
 import datetime
 from sqlalchemy import Column, Date, ForeignKeyConstraint, Index, String, Table, Text, text
-from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -11,9 +11,9 @@ class Sijoituspaikka(Base):
     __tablename__ = 'sijoituspaikka'
     __table_args__ = (ForeignKeyConstraint(['osaston_numero'], ['osastopaikka.osaston_numero'], name='sijoituspaikka_ibfk_1'), Index('IDX_Sijoituspaikka1', 'osaston_numero'))
     # Placement location number
-    sijoituspaikan_nro: Mapped[int] = mapped_column(INTEGER(11), primary_key=True)
+    sijoituspaikan_nro: Mapped[int] = mapped_column(Integer, primary_key=True)
     # Section number
-    osaston_numero: Mapped[int] = mapped_column(INTEGER(11), nullable=False, server_default=text('0'))
+    osaston_numero: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text('0'))
     # Placement date
     sijoituspvm: Mapped[Optional[str]] = mapped_column(String(255))
     # Plot
